+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/XmlFrameworkExtensionTest.php
@@ -45,7 +45,7 @@ class XmlFrameworkExtensionTest extends FrameworkExtensionTestCase
         $container = $this->createContainerFromFile('asset_mapper');
 
         $definition = $container->getDefinition('asset_mapper.public_assets_path_resolver');
-        $this->assertSame('/assets_path/', $definition->getArgument(1));
+        $this->assertSame('/assets_path/', $definition->getArgument(0));
 
         $definition = $container->getDefinition('asset_mapper.dev_server_subscriber');
         $this->assertSame(['zip' => 'application/zip'], $definition->getArgument(2));