+++ src/Symfony/Component/Notifier/Bridge/Novu/Tests/NovuOptionsTest.php
@@ -26,6 +26,12 @@ class NovuOptionsTest extends TestCase
             null,
             null,
             null,
+            [
+                'email' => [
+                    'from' => 'no-reply@example.com',
+                    'senderName' => 'No-Reply',
+                ],
+            ],
             [],
         );
 
@@ -37,6 +43,12 @@ class NovuOptionsTest extends TestCase
                 'phone' => null,
                 'avatar' => null,
                 'locale' => null,
+                'overrides' => [
+                    'email' => [
+                        'from' => 'no-reply@example.com',
+                        'senderName' => 'No-Reply',
+                    ],
+                ],
             ],
             $options->toArray()
         );