+++ src/Symfony/Component/Validator/Constraints/GroupSequenceProvider.php
@@ -11,6 +11,9 @@
 
 namespace Symfony\Component\Validator\Constraints;
 
+use Doctrine\Common\Annotations\Annotation\NamedArgumentConstructor;
+use Symfony\Component\Validator\Attribute\HasNamedArguments;
+
 /**
  * Attribute to define a group sequence provider.
  *
@@ -19,4 +22,8 @@ namespace Symfony\Component\Validator\Constraints;
 #[\Attribute(\Attribute::TARGET_CLASS)]
 class GroupSequenceProvider
 {
+    #[HasNamedArguments]
+    public function __construct(public ?string $provider = null)
+    {
+    }
 }