+++ src/Symfony/Component/Validator/Tests/Validator/RecursiveValidatorTest.php
@@ -12,6 +12,7 @@
 namespace Symfony\Component\Validator\Tests\Validator;
 
 use PHPUnit\Framework\TestCase;
+use Psr\Container\ContainerInterface;
 use Symfony\Component\Translation\IdentityTranslator;
 use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\Constraints\All;
@@ -44,6 +45,8 @@ use Symfony\Component\Validator\Mapping\Factory\MetadataFactoryInterface;
 use Symfony\Component\Validator\ObjectInitializerInterface;
 use Symfony\Component\Validator\Tests\Constraints\Fixtures\ChildA;
 use Symfony\Component\Validator\Tests\Constraints\Fixtures\ChildB;
+use Symfony\Component\Validator\Tests\Dummy\DummyGroupProvider;
+use Symfony\Component\Validator\Tests\Fixtures\Attribute\GroupProviderDto;
 use Symfony\Component\Validator\Tests\Fixtures\CascadedChild;
 use Symfony\Component\Validator\Tests\Fixtures\CascadingEntity;
 use Symfony\Component\Validator\Tests\Fixtures\EntityWithGroupedConstraintOnMethods;
@@ -57,6 +60,7 @@ use Symfony\Component\Validator\Validator\ContextualValidatorInterface;
 use Symfony\Component\Validator\Validator\LazyProperty;
 use Symfony\Component\Validator\Validator\RecursiveValidator;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
+use Symfony\Contracts\Service\ServiceLocatorTrait;
 
 class RecursiveValidatorTest extends TestCase
 {
@@ -1262,6 +1266,22 @@ class RecursiveValidatorTest extends TestCase
         }
     }
 
+    public function testGroupProvider()
+    {
+        $dto = new GroupProviderDto();
+
+        $metadata = new ClassMetadata($dto::class);
+        $metadata->addPropertyConstraint('firstName', new NotBlank(groups: ['foo']));
+        $metadata->addPropertyConstraint('lastName', new NotBlank(groups: ['foo']));
+        $metadata->setGroupProvider(DummyGroupProvider::class);
+        $metadata->setGroupSequenceProvider(true);
+        $this->metadataFactory->addMetadata($metadata);
+
+        $violations = $this->validate($dto, null, 'Default');
+
+        $this->assertCount(2, $violations);
+    }
+
     public static function getConstraintMethods()
     {
         return [
@@ -2040,8 +2060,14 @@ class RecursiveValidatorTest extends TestCase
 
         $contextFactory = new ExecutionContextFactory($translator);
         $validatorFactory = new ConstraintValidatorFactory();
+        $factories = [
+            DummyGroupProvider::class => static fn () => new DummyGroupProvider(),
+        ];
+        $groupProviderLocator = new class($factories) implements ContainerInterface {
+            use ServiceLocatorTrait;
+        };
 
-        return new RecursiveValidator($contextFactory, $metadataFactory, $validatorFactory, $objectInitializers);
+        return new RecursiveValidator($contextFactory, $metadataFactory, $validatorFactory, $objectInitializers, $groupProviderLocator);
     }
 
     public function testEmptyGroupsArrayDoesNotTriggerDeprecation()